add legacy switching logic
