+++ assets/scss/builder/panel/_advanced.scss
@@ -1,4 +1,4 @@
-.editor-panel[data-type="background"],
+.editor-panel[data-type="background-legacy"],
 .editor-panel[data-type="Advanced"] {
 	.customize-navigation[data-element-type="table-cell"] {
 		.item[data-control-name="margin"],
