+++ assets/scss/builder/panel/_background.scss
@@ -1,5 +1,5 @@
 /*************** Background Panel ********************/
-.editor-panel[data-type="background"] {
+.editor-panel[data-type="background-legacy"] {
 	img {
 		max-width: 100%;
 	}
@@ -294,6 +294,29 @@
 	.panel-footer {
 		.customize-design {
 			display: none;
+		}
+	}
+}
+
+.editor-panel[data-type="background"] {
+	.background-panel-section {
+		display: none;
+		&.active {
+			display: block;
+		}
+	}
+	.background-nav {
+		display: block;
+		.background-nav-item {
+			width: 60px;
+			span.dashicons {
+				color: white;
+				font-size:30px;
+				text-align: center;
+				width: 60px;
+				height: 40px;
+				padding-top: 10px;
+			}
 		}
 	}
 }
