+++ assets/scss/builder/panel/_navigation.scss
@@ -9,7 +9,7 @@
 	}
 }
 
-.customize-navigation {
+.customize-navigation, .background-nav {
 	position: absolute;
 	bottom: 0;
 	left: 0;
