+++ package.json
@@ -110,7 +110,8 @@
 		"webpack-cli": "^2.0.12",
 		"webpack-dev-server": "^3.11.2",
 		"wowjs": "^1.1.3",
-		"wp-color-picker-alpha": "git+https://github.com/rramo012/wp-color-picker-alpha.git#master"
+		"wp-color-picker-alpha": "git+https://github.com/rramo012/wp-color-picker-alpha.git#master",
+		"semver": "^7.3.8"
 	},
 	"greenkeeper": {
 		"ignore": [
